+++ cross/applications/pages/steps/categorical_encoding/categorical_encoding_page.py
@@ -104,7 +104,7 @@ class CategoricalEncodingPage(CategoricalEncodingBase):
             # Determine if new columns are created
             new_columns = (
                 list(set(transformed_df.columns) - set(original_df.columns))
-                if encoding_type in ["onehot", "dummy", "binary"]
+                if encoding_type in ["onehot", "dummy", "binary", "target"]
                 else [column]
             )