+++ cross/transformations/feature_engineering/categorical_enconding.py
@@ -1,7 +1,12 @@
 import pandas as pd
-from category_encoders import BinaryEncoder, TargetEncoder
+from category_encoders import BinaryEncoder
 from sklearn.base import BaseEstimator, TransformerMixin
-from sklearn.preprocessing import LabelEncoder, OneHotEncoder, OrdinalEncoder
+from sklearn.preprocessing import (
+    LabelEncoder,
+    OneHotEncoder,
+    OrdinalEncoder,
+    TargetEncoder,
+)
 
 
 class CategoricalEncoding(BaseEstimator, TransformerMixin):
@@ -51,7 +56,7 @@ class CategoricalEncoding(BaseEstimator, TransformerMixin):
             self._encoders[column] = BinaryEncoder().fit(X[[column]])
 
         elif transformation == "target" and y is not None:
-            self._encoders[column] = TargetEncoder().fit(X[[column]], y)
+            self._encoders[column] = TargetEncoder(smooth=0).fit(X[[column]], y)
 
         elif transformation == "count":
             self._encoders[column] = X[column].value_counts().to_dict()
@@ -78,17 +83,22 @@ class CategoricalEncoding(BaseEstimator, TransformerMixin):
                 lambda val: self._safe_transform(val, transformer, known_classes)
             )
 
-        elif transformation in ["ordinal", "target"]:
+        elif transformation in ["ordinal"]:
             X[column] = transformer.transform(X[[column]])
 
-        elif transformation in ["onehot", "dummy", "binary"]:
+        elif transformation in ["onehot", "dummy", "binary", "target"]:
             encoded_array = transformer.transform(X[[column]])
+            columns = transformer.get_feature_names_out([column])
+
+            if transformation == "target":
+                columns = [f"{col}_target" for col in columns]
+
             encoded_df = pd.DataFrame(
                 encoded_array,
-                columns=transformer.get_feature_names_out([column]),
+                columns=columns,
                 index=X.index,
             )
-            X = pd.concat([X, encoded_df], axis=1).drop(columns=[column])
+            X = pd.concat([X.drop(columns=[column]), encoded_df], axis=1)
 
         elif transformation == "count":
             X[column] = X[column].map(self._encoders[column]).fillna(0)