+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Target encoding for multiclass in CategoricalEncoding
+
 ## [1.0.0] - 2025-01-03
 
 ### Fixed